lisp/casual-compile: a port of counsel-compile for casual
counsel-compile provides a nice auto-narrowing interface to the M-x
compile command that can build directories and targets to provide a
fast interface to select different build targets.

This attempts to implement a transient-like interface to the same
functionality. We import some of the helper functions under the:

  casual-compile--

namespace and provide the final casual-compile invocation.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>

---
v2
  - use buffer local variables transient args
  - persist args via per project .dir-locals2
  - added env handling
  - use define-suffix helpers
  - general refactoring